+++ testapps/jpa-tester/src/main/java/test/jpa/AppEntry.java
@@ -1,0 +1,12 @@
+package test.jpa;
+
+import act.Act;
+
+@SuppressWarnings("unused")
+public class AppEntry {
+
+    public static void main(String[] args) throws Exception {
+        Act.start();
+    }
+
+}