+++ testapps/jpa-tester/src/main/java/test/jpa/Gender.java
@@ -1,0 +1,5 @@
+package test.jpa;
+
+public enum Gender {
+    M, F
+}